refactor: extract calendar bar scaling helpers
